+++ src/lib/simulation/manifests.ts
@@ -22,8 +22,10 @@ import { manifest as nuclearReactor } from './simulations/nuclear-reactor.manife
 import { manifest as newmanProjection } from './simulations/newman-projection.manifest';
 import { manifest as crystalField } from './simulations/crystal-field.manifest';
 import { manifest as reactionKinetics } from './simulations/reaction-kinetics.manifest';
+import { manifest as electronConfiguration } from './simulations/electron-configuration.manifest';
+import { manifest as chemicalBond } from './simulations/chemical-bond.manifest';
 
-export const MANIFESTS: SimulationManifest[] = [projectileMotion, waveSpeed, frequencyPeriod, pendulum, heatTransfer, ohmsLaw, shmSpring, idealGasLaw, momentumCollision, inclinedPlane, dopplerEffect, unitCircle, quadratic, probability, nuclearReactor, newmanProjection, crystalField, reactionKinetics];
+export const MANIFESTS: SimulationManifest[] = [projectileMotion, waveSpeed, frequencyPeriod, pendulum, heatTransfer, ohmsLaw, shmSpring, idealGasLaw, momentumCollision, inclinedPlane, dopplerEffect, unitCircle, quadratic, probability, nuclearReactor, newmanProjection, crystalField, reactionKinetics, electronConfiguration, chemicalBond];
 
 export const manifestBySlug = new Map(MANIFESTS.map((m) => [m.metadata.slug, m]));
 export const manifestByProcessorId = new Map(MANIFESTS.map((m) => [m.metadata.processorId, m]));